# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Your API"
  description: "This is an API"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /customers/login:
    post:
      tags:
        - Customers
      summary: customers login
      description: customers login credentials using email and password
      parameters:
      - in: body
        name: body
        description: customer login credentials
        required: true
        schema:
          type: object
          properties:
            email:
              type: string
            password:
              type: string 
      responses:
        201:
          description: customer login successful
          schema:
            type: object
            properties:
              token:
                type: string
          examples:
            application/json:
              token: "your_token_here"
  /customers:
    post:
      tags:
        - Customers
      summary: Create a new customer
      description: Info to create the customer
      parameters:
        - in: body
          name: body
          description: Customer information
          required: true
          schema:
            $ref: '#/definitions/CustomerPayload'
      responses:
        201:
          description: customer created successfully
          schema:
            $ref: '#/definitions/CustomerResponse'
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
            examples:
              application/json:
                message: "Invalid input"

    put:
      tags:
        - Customers
      summary: update customer
      description: updating customer information
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: Customer information used for update
          required: true
          schema:
            $ref: '#/definitions/CustomerPayload'
      responses:
        201:
          description: customer updated successfully
          schema:
            $ref: '#/definitions/CustomerResponse'
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
            examples:
              application/json:
                message: "Invalid input"
      
    delete:
      tags:
        - Customers
      summary: "Enpoint to delete a customer"
      description: "Endpoint to delete a customer. Requires token authentication, member must be logged in to delete their account"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully Deleted member"           
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "successfully deleted customer."
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Get a customer by ID
      description: Retrieve a customer by their ID
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: ID of the customer to retrieve
      responses:
        200:
          description: Customer retrieved successfully
          schema:
            $ref: '#/definitions/CustomerResponse'
  /customers/paginate?page={page}&per_page={per_page}:
    get:
      tags:
        - Customers
      summary: Get all customers with pagination
      description: Retrieve all customers with pagination
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
          description: Page number for pagination
        - name: per_page
          in: query
          required: true
          schema:
            type: integer
          description: Number of customers per page
      responses:
        200:
          description: Customers retrieved successfully
          schema:
            type: object
            properties:
              total_customers:
                type: integer
              page:
                type: integer
              per_page:
                type: integer
              customers:
                type: array
                items:
                  $ref: '#/definitions/CustomerListResponse'
          examples:
            application/json:
              total_customers: 1
              page: 1
              per_page: 1
              customers:
                - email: "kaliuu@email.com"
                  id: 1
                  name: "Kali Uchi"
                  password: "scrypt:32768:8:1$C1paC1ITGzKkKwz3$0a18aeffee56368f8f66a6684ab93f7e6dc7ab1161d1b1e5fead519fda9e79ca32dbf238eedbbe4282a83fe07c2c3b9dbf7d4fed5a93deb7e2f7f0bcb73f5b6a"
                  phone: "123-678-2223"
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: Get all tickets for a customer
      description: Retrieve all tickets for a customer
      security:
        - bearerAuth: []
      responses:
        200:
          description: Tickets retrieved successfully
          schema:
            type: object
            properties:
              tickets:
                type: array
                items:
                  type: integer
            examples:
              application/json:
                tickets: [1, 2, 3]
  /customers/search:
    get:
      tags:
        - Customers
      summary: Search for customers
      description: Search for customers by name
      parameters:
        - name: search
          in: query
          required: true
          type: string
          description: Search query (name)
      responses:
        200:
          description: Customers retrieved successfully
          schema:
            type: object
            properties:
              customers:
                type: array
                items:
                  $ref: '#/definitions/CustomerListResponse'
# -------------------------------------------------------------------------=
  /inventory/paginate?page={page}&per_page={per_page}:
    get:
      tags:
        - Inventory
      summary: Get all inventory items with pagination
      description: Retrieve all inventory items with pagination
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
          description: Page number for pagination
        - name: per_page
          in: query
          required: true
          schema:
            type: integer
          description: Number of inventory items per page
      responses:
        200:
          description: Inventory items retrieved successfully
          schema:
            type: object
            properties:
              total_items:
                type: integer
              page:
                type: integer
              per_page:
                type: integer
              items:
                type: array
                items:
                  $ref: '#/definitions/InventoryListResponse'
  /inventory:
    post:
      tags:
        - Inventory
      summary: Create a new inventory item
      description: Info to create the inventory item
      parameters:
        - in: body
          name: body
          description: Inventory item information
          required: true
          schema:
            type: object
            properties:
              part_name:
                type: string
              brand:
                type: string
              price:
                type: integer
      responses:
        201:
          description: Inventory item created successfully
          schema:
            $ref: '#/definitions/InventoryListResponse'
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
            examples:
              application/json:
                message: "Invalid input"
  /inventory/{inventory_id}:
    get:
      tags:
        - Inventory
      summary: Get an inventory item by ID
      description: Retrieve an inventory item by its ID
      parameters:
        - name: inventory_id
          in: path
          required: true
          type: integer
          description: ID of the inventory item to retrieve
      responses:
        200:
          description: Inventory item retrieved successfully
          schema:
            $ref: '#/definitions/InventoryListResponse'

    put:
      tags:
        - Inventory
      summary: Update an inventory item
      description: Update an inventory item by its ID
      parameters:
        - name: inventory_id
          in: path
          required: true
          type: integer
          description: ID of the inventory item to update
        - in: body
          name: body
          description: Inventory item information used for update
          required: true
          schema:
            type: object
            properties:
              part_name:
                type: string
              brand:
                type: string
              price:
                type: integer
      responses:
        200:
          description: Inventory item updated successfully
          schema:
            $ref: '#/definitions/InventoryListResponse'
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
            examples:
              application/json:
                message: "Invalid input"
              
    delete:
      tags:
        - Inventory
      summary: Delete an inventory item by ID
      description: Delete an inventory item by its ID
      parameters:
        - name: inventory_id
          in: path
          required: true
          type: integer
          description: ID of the inventory item to delete
      responses:
        200:
          description: Inventory item deleted successfully
          schema:
            $ref: '#/definitions/DeleteInventoryResponse'
          examples:
            application/json:
              message: "successfully deleted inventory item."
# ----------------------------------------------------------------------
...
definitions:
...
  DeleteInventoryResponse:
    type: object
    properties:
      message:
        type: string
  InventoryListResponse:
    type: array
    items:
      type: object
      properties:
        id:
          type: integer
        part_name:
          type: string
        brand:
          type: string
        price:
          type: integer
  DeleteCustomerResponse:
    type: object
    properties:
      message:
        type: string
  CustomerPayload:
    type: object
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
  CustomerResponse:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
  CustomerListResponse:
    type: array
    items:
      type: object
      properties:
        email:
          type: string
        id:
          type: integer
        name:
          type: string
        password:
          type: string
        phone:
          type: string
...